# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Documents Service API
  version: 0.1.0
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    UserIdHeader:
      type: apiKey
      in: header
      name: X-User-Id
  schemas:
    Folder:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        parentFolderId: { type: string, format: uuid, nullable: true }
        propertyId: { type: string, format: uuid, nullable: true }
        transactionId: { type: string, format: uuid, nullable: true }
        transactionStage: { type: string, nullable: true }
        ownerUserId: { type: string }
...
security:
  - UserIdHeader: []
paths:
...
  /folders:
    post:
      summary: Create folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                parentFolderId: { type: string, format: uuid, nullable: true }
                propertyId: { type: string, format: uuid, nullable: true }
                transactionId: { type: string, format: uuid, nullable: true }
                transactionStage: { type: string, nullable: true }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Folder' }
    get:
      summary: List folders
      parameters:
        - in: query
          name: parentFolderId
          schema: { type: string, format: uuid }
        - in: query
          name: transactionId
          schema: { type: string, format: uuid }
        - in: query
          name: propertyId
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
  /folders/{folderId}:
    get:
      summary: Get folder
      parameters:
        - in: path
          name: folderId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
    patch:
      summary: Update folder
      parameters:
        - in: path
          name: folderId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
    delete:
      summary: Delete folder
      parameters:
        - in: path
          name: folderId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
  /folders/{folderId}/permissions:
    get:
      summary: List folder sharing permissions
      parameters:
        - in: path
          name: folderId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
    post:
      summary: Invite collaborator to folder
      parameters:
        - in: path
          name: folderId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201': { description: Created }
...